requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
easy-pil==0.1.9
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
Pillow==8.4.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
